requirements.txt
requests
charset-normalizer>=2,<4
python-dotenv
APScheduler
pytz
tzdata